+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Sin Publicar
 - Added Multiple Row Component
 - Resized pill icon on Hybrid Carousel
+- Fixed Row Component
 
 # v1.18.0
 🚀 1.18.0 🚀