+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Resized pill icon on Hybrid Carousel
 - Added Multiple Description Component
 - Added Multiple Description to Row
+- New image accesory for Hyrbrid Carousel Card
 
 # v1.18.0
 🚀 1.18.0 🚀